Test Securty Of Review  Module

+++ app/Http/Controllers/Api/ReviewController.php
@@ -7,26 +7,19 @@ use App\Http\Requests\Review\StoreReviewRequest;
 use App\Http\Requests\Review\UpdateReviewRequest;
 use App\Models\Review;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class ReviewController extends Controller
 {
     /**
-     * Create a new AuthController instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->middleware('admin', ['except' => ['index','show']]);
-    }
-
-    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * 
+     * Just For auth user 
      */
-    public function index()
+    public function index() // Secured Endpoint 
     {
         $reviews = DB::table('reviews')->paginate(5);
         
@@ -41,8 +34,10 @@ class ReviewController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
+     * 
+     * Just For auth user
      */
-    public function store(StoreReviewRequest $request)
+    public function store(StoreReviewRequest $request) // Secured Endpoint 
     {
         $request = $request->validated();
 
@@ -56,10 +51,12 @@ class ReviewController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  Review $review
      * @return \Illuminate\Http\Response
+     * 
+     * Just For auth user
      */
-    public function show(Review $review)
+    public function show(Review $review) // Secured Endpoint 
     {
         return response()->json([
             'success' => true,
@@ -71,11 +68,21 @@ class ReviewController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  Review $review
      * @return \Illuminate\Http\Response
+     * 
+     * For Specific User , who created this rec
+     * 
      */
-    public function update(UpdateReviewRequest $request, Review $review)
+    public function update(UpdateReviewRequest $request, Review $review) // Secured Endpoint 
     {
+        if ($review->user_id != Auth::user()->id ) 
+        {
+            return response()->json([
+                'message' => "Unauthenticated, this action for specific user only",
+            ],401);
+        }
+
         $request = $request->validated();
 
         $review = $review->update( $request );
@@ -88,11 +95,21 @@ class ReviewController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  Review $review
      * @return \Illuminate\Http\Response
+     * 
+     * For Specific User , who created this rec
+     * 
      */
-    public function destroy(Review $review)
+    public function destroy(Review $review) // Secured Endpoint 
     {
+        if ($review->user_id != Auth::user()->id ) 
+        {
+            return response()->json([
+                'message' => "Unauthenticated, this action for specific user only",
+            ],401);
+        }
+
         $review = $review->delete( $review );
 
         if ($review) return response()->json([
